Rename index.html, code.js, style.css to wifi.{html,js,css}
When this library includes static assets as embedded files in
CMakeLists.txt called index.html, code.css, style.css, it gets exclusive
access to files of that name.

This means that when projects include this library, they cannot embed
files with the same names.

This change alters the library to instead defined the files wifi.html,
wifi.js, and wifi.css so that the original names are available to
parent projects.